requirements.txt
discord.py==2.3.2
telethon==1.28.4
aiohttp==3.8.6
python-dotenv==1.0.0